+++ packages/mcp-http-fetch/CHANGELOG.md
@@ -4,4 +4,7 @@
 
 ### Minor Changes
 
-- [#4](https://github.com/ver0-project/mcps/pull/4) [`5a6d19a`](https://github.com/ver0-project/mcps/commit/5a6d19ab3caf7c9bda7bb3392d8b15c59e44f02e) Thanks [@xobotyi](https://github.com/xobotyi)! - Initial implementation of package with fetch-html, fetch-markdown and fetch-json tools
+- [#4](https://github.com/ver0-project/mcps/pull/4)
+  [`5a6d19a`](https://github.com/ver0-project/mcps/commit/5a6d19ab3caf7c9bda7bb3392d8b15c59e44f02e) Thanks
+  [@xobotyi](https://github.com/xobotyi)! - Initial implementation of package with fetch-html, fetch-markdown and
+  fetch-json tools